server/core/components/workers/requirements.txt
cryptography
pymysql
pyOpenSSL
redis
python-etcd
